+++ Editor/Source/EditorLayer.h
@@ -4,9 +4,15 @@
 
 namespace Core
 {
+    struct EditorState
+    {
+    };
+
     class EditorLayer : public Layer
     {
     public:
+        EditorState state;
+
         EditorLayer() {};
         ~EditorLayer() {};